Add getaddrinfo() and freeaddrinfo().
getaddrinfo() really just calls into the middleware to get an array of services
and an array of addresses, then returns their cross product.

As a side effect, looking up a name reveals the canonical name. As part of the
contract, I will have to copy the input name into the canonical name and get it
overwritten by the end. Oh, we will see.

The memory management could be complicated if I actually implemented it with
linked lists and everything in its own little block. But I will instead put
everything into the same allocation. freeaddrinfo() then can free everything in
one go if it is called on the first element. It can also be called on any other
element, in which case it does nothing.

+++ src/include/netdb.h
@@ -3,6 +3,7 @@
 
 #include "../../include/netdb.h"
 
+#include <stddef.h>
 #include <stdint.h>
 
 /* getaddrinfo()/gethostbyname()/getservbyname() middleware */
@@ -28,7 +29,7 @@ extern hidden int __lookup_ipliteral(const char *, int, struct address *);
 #define MAXNS           3
 struct resolvconf {
     struct address ns[MAXNS];
-    size_t nns;
+    int nns;
     char search[256];
     int ndots;
     int timeout;
